captcha_eval_framework/requirements.txt
openai
playwright
Pillow
tqdm
